requirements.txt
numpy
scikit-learn
python-dotenv

langchain
langchain-core
langchain-community

langchain-openai
langchain-ollama

# visualize chains
grandalf

# for web loader
bs4

# for pdf loader
pypdf